update flannle install method to helm

+++ applications/helmfile/latest/scripts/disable.sh
@@ -7,11 +7,16 @@ function log::info() {
   printf "%s \033[32minfo \033[0m%s\n" "$(date +'%Y-%m-%dT%H:%M:%S')" "$*"
 }
 
-app_name="helmfile"
-log::info "uninstalling ${app_name}"
+NAME="helmfile"
 sealos exec -r master "
+binary=/usr/local/bin/helmfile
+if ! [ -x $binary ]; then
+  echo "Helmfile is not installed"
+  exit 0
+fi
+echo 'Removing Helmfile ...'
 rm -rf /root/.local/share/helm/plugins/helm-diff
 rm -rf /usr/bin/helm
 rm -rf /usr/bin/helmfile
 "
-log::info "${app_name} is uninstalled"
+log::info "${NAME} has been removed"
